+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - French translations for the plugin interface
 - Translation support with textdomain loading
 
+### Fixed
+- Time zone issue where post idea creation times were displayed in GMT instead of local time (France)
+
 ## [1.0.0] - 2024-06-01
 
 ### Added